Create modulusThe ifelse Statement
The if statement is a one-way choice: if the condition is true we do something, if not we don't. We would now like an either/or, or two-way choice: if the condition is true, do this, if not, do that. Once we can accomplish that, we can respond one way if the user is under age, and another way if the user is 21 or greater.

For this we add an else statement:

#include <iostream>
#include <string>
using namespace std;
 
int main()
{ 
   int age;

   cout << "How old are you? ";
   cin >> age;

   if (age >= 21)
      cout << "Would you like a beer?\n";
   else
      cout << "Sorry, I can't serve you.\n";   
   cout <<"Have a nice day.\n\n";   
}
Try the program after making this change and see how it behaves.

Here's another example.  Remember how to generate a remainder? Use the modulo operator, %. Here is a console application that gets a number from the user and tells the user whether or not the number is even or odd.  Make sure you can follow the logic - why does this program give the correct answer?  You have to understand the % operator to answer that question.

#include <iostream>
#include <string>
using namespace std;
 
int main()
{ 
   int userValue;

   cout << "Enter an int: ";

   // get the user's number
   cin >> userValue;

   if ( userValue % 2 == 0 )
      // if we divide by 2 with no remainder, it's even
      cout << userValue << " is even.\n\n";
   else
      // if we divide by 2 with a 1 remainder, it's odd
      cout << userValue << " is odd.\n\n";
}