Create the entire else ifchain:
f/else ... if/else  ...
A common control structure in C++ is one in which a number of mutually exclusive conditions are expected, each with its own consequences. By inserting else if between an if statement and an else statement, we get this effect — as soon as a condition in the if or one of the if/else lines is met, the statement or block it controls is executed, and the program continues by skipping to the end of the entire else ifchain:

#include <iostream>
#include <string>
using namespace std;
 
int main()
{ 
   int depth;

   cout << "What is your diving depth in feet? ";
   cin >> depth;

   if (depth < 33)
      cout <<  "Use 2 atm for your calculation.\n";
   else if (depth < 66)
      cout <<  "Use 3 atm for your calculation.\n";
   else if (depth < 99)
      cout <<  "Use 4 atm for your calculation.\n";
   else if (depth < 132)
      cout <<  "Use 5 atm for your calculation.\n";
   else
      cout <<  "You are down too deep!\n";
}
It looks as if they are not really mutually exclusive, since the < 33 case is conceptually covered by the < 66 case. But since we don't test < 66 unless the < 33 has failed, then the < 66 case is really going to catch numbers between 33 and 66, and not < 33.

As a technical matter, these are really a bunch of nested if/else statements (but don't indent like this):

      if (depth < 33)
         cout << "Use 2 atm for you calculation. \n";
      else
      {
         if (depth < 66)
            cout << "Use 3 atm for you calculation. \n";
         else
         {
            if (depth < 99)
               cout << "Use 4 atm for you calculation. \n";
            else 
            {
               if (depth < 132)
                  cout << "Use 5 atm for you calculation. \n";
               else
                  cout << "You are way too deep! \n";
            }
         }
      }
Showing this nesting does nothing to improve the understanding of the program. In fact, it obfuscates the true nature of our logic, which is to pigeon-hole a value. The above poor style, is called progressive indentation. It should not be used if all the various cases being tested have equal status. There is nothing "better" about the depth being in the 66-99 range than there is about it being in 33-66 range.

If there is logic that is truly nested, we would use progressive indentation and perhaps braces, even when not needed:

      if (state == 'y')