Create logicaloperatorsagain
Logical Operators Again
Logical Operators
If the user types a capital Y instead of a lower-case y in response to:

     cout << "Are you in-state? ";
then the test:

     if (state == 'y')
would be considered false. We want to allow for either a Y or a y. This can be done by combining the two relationals:

    (state == 'y')
and

    (state == 'Y')
using the logical or, ||, like so:

     if ( (state == 'y') || (state == 'Y') )
(To type the || you would hit the vertical pipe bar, |, twice.)

Now the entire expression will evaluate to true if either state=='y' OR state=='Y'. The parentheses are not necessary in this expression, but they are a good idea: C++ has lots of operators whose precedence is counterintuitive. Instead of trying to remember which they are, just over-do the parentheses.

Truth Tables for the Three Principal Logical Operators
The or, || operator, and, &&, operator and not, !, operator are defined like so:

OR
a	b	a||b
true	true	true
false	true	true
true	false	true
false	false	false
AND
a	b	a&&b
true	true	true
false	true	false
true	false	false
false	false	false
NOT
a	!a
true	false
false	true
With these statements, you can build rather complex logical expressions. You will see and create many, and there are lots of examples in the text. For now the important thing to remember is that whenever a logical expression is called for, it evaluates to either true or false, and this is interpreted by the larger statement (the if statement, e.g.) and a certain action is taken as a result.

An Example
Here is a (silly but illustrative) state tax example.

Pay attention to:
Indentation.
User Input
Conversion of string to char
Logical expression.
#include <iostream>
#include <string>
using namespace std;
 
int main()
{ 
   char inState;
   double price, tax;
   string tempString;

   // find out if we should charge state tax          
   cout << "Are you in-state? ";      
   getline(cin, tempString);
   
   // copy the first char only
   inState = tempString[0]; 

   // Get the amount          
   cout << "What is the amount? ";      
   cin >> price;

   if (inState == 'y' || inState == 'Y')
   {
      if (price < .12)
        tax = 0;
      else if (price < .24)
        tax = .01;
      else if (price < .36)
        tax = .02;
      else
        tax = price * .083;
   }
   else
      tax = 0;
       
   cout << "The amount of tax is " << tax << endl; 
}