Create if ( userString == "Yes" )
An Example with Strings
If we want to test whether the user typed a "yes" or "no" in response to a question, one easy way to do it (as we have already demonstrated) is to pluck the first character from the answer and test it against 'y' or 'n'.  This gives the user a little leeway in his spelling skills.  If he messes up the rest of the word, we will forgive him. As long as he types the first letter correctly, we'll act on that, alone.

What if we want to make sure the user types some word or phrase, exactly?  We have to compare the entire string.  The way to do this in C++ is using the == operator. It is very simple:

   if ( userString == "Yes" )
      cout << "You typed Yes";
We can turn this into a complete program that demands that the user type in either the word "Yes" or "yes", but nothing else.

#include <iostream>
#include <string>
#include <sstream>
using namespace std;
using namespace std;
int main()
{ 
   char inState;
   double price, tax;
   string userString;

   // grab the string          
   cout << "Type in a word: ";      
   getline(cin, userString);

   if (userString == "yes" || userString == "Yes")
      cout << "You typed an acceptable form of \"yes\".";
   else
      cout << "You did not type an acceptable form of \"yes\".";
}

/* ------------------- Sample Run #1 --------------------
Type in a word: YES
You did not type an acceptable form of "yes".
Press any key to continue . . .
---------------------- End Sample Run #1 ---------------- */


/* ------------------- Sample Run #2 --------------------
Type in a word: Yes
You typed an acceptable form of "yes".
Press any key to continue . . .
---------------------- End Sample Run #2 ---------------- */
Notice that we have also demonstrated a technique for including a double quote mark, ", inside a string literal by preceding it with a so-called escape character, otherwise known as the backslash, \.

Warning About Logical Operators
Be careful to use double && and ||, not single & and |.  In C++ the single operators mean something entirely different than the double operators and will usually result in logical errors (that means bugs).