Create limerickexcercise
how are the variables returned_guess and guess connected in the program ....#include <iostream>
#include <string>

using namespace std;

string limerick = "a limerick has five lines,\nwith two feet in the fourth,\nand two in the third,\nthe second has three,\nand you can tell it to a _______.";
string limerick1 ="a limerick has five lines,\nwith two feet in the fourth,\nand two in the third,\nthe second has three,\nand you can tell it to a __tree_____.";
string last_word = "tree";

string GuessLimerick() {
 string guess;
 int chances = 5;
 int i = 0;

 cout << limerick;
 cout << "What is the last word in the limerick? (use lower case letters)\n";
 cin >> guess;

 while (guess != last_word && i < chances) {
  cout << "Incorrect. You have " << chances - i - 1 << " chances left.\n";
  cin >> guess;
  i++;
 }

 if (guess == last_word) {
 cout << "Correct!\n";
  
 } else {
 cout << "Sorry, you ran out of chances. The last word is " << last_word << ".\n";
 }

 return guess;
}

int main() {
 string choice;
 string returned_guess;

 while (choice != "exit") {
 cout << "menu:\n";
 cout << "play\n";
 cout << "cheat\n";
 cout << "exit\n";
 cout << "Enter your choice: ";
 cin >> choice;

 if (choice == "play") {
  returned_guess = GuessLimerick();
  if (returned_guess == last_word) { // If the user guessed the last word, end the program.
   break;
  }
 } else if (choice == "cheat") {
  cout << limerick1 << endl;
  break; // End the program after printing the limerick.
 } else if (choice == "exit") {
  break; // End the program if the user chooses to exit.
 }
 }

 return 0;
}
/*menu:[OUTPUT]
play
cheat
exit
Enter your choice: play
a limerick has five lines,
with two feet in the fourth,
and two in the third,
the second has three,
and you can tell it to a _______.What is the last word in the limerick? (use lower case letters)
feet
Incorrect. You have 4 chances left.
tree
Correct!*/
View other drafts





















The variables returned_guess and guess are connected in the program in the following way:

guess is the variable that stores the user's guess for the last word in the limerick.
returned_guess is the variable that stores the returned value of the GuessLimerick() function.
The GuessLimerick() function takes no arguments and returns a string. The string that is returned is the user's guess for the last word in the limerick, or the empty string if the user ran out of chances to guess correctly.

In the main() function, the returned_guess variable is assigned the value of the GuessLimerick() function. This means that the returned_guess variable will contain the user's guess for the last word in the limerick, or the empty string if the user ran out of chances to guess correctly.

The returned_guess variable is then used in the following way:

If the returned_guess variable is equal to the last word in the limerick, then the program breaks out of the loop and ends.
If the returned_guess variable is not equal to the last word in the limerick, then the program continues to loop and asks the user to guess again.
Here is an example of how the returned_guess variable is used in the program:

C++
if (returned_guess == last_word) {
  // The user guessed the last word correctly.
  break;
} else {
  // The user did not guess the last word correctly.
  // Continue to loop and ask the user to guess again.
}